moved command to command folder
